Move OpenURLInBrowser to shared package
